+++ src/index.ts
@@ -1,7 +1,0 @@
-const prettierIgnore = require('./.prettierIgnore.js');
-
-import prettierConfig from './prettier.config';
-export default {
-  prettierConfig,
-  prettierIgnore,
-};