+++ CHANGELOG.md
@@ -1,7 +1,54 @@
+## [1.1.0-dev.1](https://github.com/SebastianWesolowski/s-prettier/compare/v1.0.0...v1.1.0-dev.1) (2024-02-11)
+
+
+### Features
+
+* ✨ added default list for prettier ignore ([59f5880](https://github.com/SebastianWesolowski/s-prettier/commit/59f58804f5596653a4ff50cb4d1002a2c5a73d1f))
+
+
+### Bug Fixes
+
+* 🐛 overridden default configuration option removed ([2ff0b20](https://github.com/SebastianWesolowski/s-prettier/commit/2ff0b2093e3278a624fc13ef36eb5de3d25bfe30))
+* 🐛 overridden default configuration option removed ([fdbbf3f](https://github.com/SebastianWesolowski/s-prettier/commit/fdbbf3f84abbe091422e985014f043f0cfa8e453))
+* 🐛 use correct config file to export ([cd41d38](https://github.com/SebastianWesolowski/s-prettier/commit/cd41d38d47f604a77e4d234d4312f638869d9e0c))
+* 🐛 use singleQuote ([d7e1438](https://github.com/SebastianWesolowski/s-prettier/commit/d7e1438a73c9cb56d2e604e5df8b2499488b8217))
+
+## [1.0.0-dev.5](https://github.com/SebastianWesolowski/s-prettier/compare/v1.0.0-dev.4...v1.0.0-dev.5) (2024-02-11)
+
+
+### Bug Fixes
+
+* 🐛 use singleQuote ([d7e1438](https://github.com/SebastianWesolowski/s-prettier/commit/d7e1438a73c9cb56d2e604e5df8b2499488b8217))
+
+## [1.0.0-dev.4](https://github.com/SebastianWesolowski/s-prettier/compare/v1.0.0-dev.3...v1.0.0-dev.4) (2024-02-11)
+
+
+### Bug Fixes
+
+* 🐛 use correct config file to export ([cd41d38](https://github.com/SebastianWesolowski/s-prettier/commit/cd41d38d47f604a77e4d234d4312f638869d9e0c))
+
+## [1.0.0-dev.3](https://github.com/SebastianWesolowski/s-prettier/compare/v1.0.0-dev.2...v1.0.0-dev.3) (2024-02-10)
+
+
+### Bug Fixes
+
+* 🐛 overridden default configuration option removed ([2ff0b20](https://github.com/SebastianWesolowski/s-prettier/commit/2ff0b2093e3278a624fc13ef36eb5de3d25bfe30))
+* 🐛 overridden default configuration option removed ([fdbbf3f](https://github.com/SebastianWesolowski/s-prettier/commit/fdbbf3f84abbe091422e985014f043f0cfa8e453))
+
+## [1.0.0-dev.2](https://github.com/SebastianWesolowski/s-prettier/compare/v1.0.0-dev.1...v1.0.0-dev.2) (2024-02-07)
+
+
+### Features
+
+* ✨ added default list for prettier ignore ([59f5880](https://github.com/SebastianWesolowski/s-prettier/commit/59f58804f5596653a4ff50cb4d1002a2c5a73d1f))
+
+## 1.0.0-dev.1 (2024-02-06)
+
+
 ## 1.0.0 (2024-02-05)
 
 
 ### Features
 
-* ✨ expor initial configuration ([f80734f](https://github.com/SebastianWesolowski/s-prettier/commit/f80734f83a842f17b6ec1be8b2bd2044a51e4c4a))
+* ✨ export initial configuration ([f80734f](https://github.com/SebastianWesolowski/s-prettier/commit/f80734f83a842f17b6ec1be8b2bd2044a51e4c4a))
 * ✨ export pretteir configs ([705da82](https://github.com/SebastianWesolowski/s-prettier/commit/705da828f57f510b2e32fd59a5a9164c5b9e616a))